+++ src/main/java/com/bad/bigad/service/game/GameRoomService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface GameRoomService {
     public GameRoom getGameRoomById(long id);
-    public GameRoom newGameRoom(int mapId, int roomType);
+    public GameRoom newGameRoom(long mapId, int roomType);
     public List<GameRoom> getAllGameRoom();
     public void updateGameRoom(GameRoom gameMap);
 }
